Added Assimp ModelLoader and read_obj has been removed. Now using interleaved vertex buffers. Renderer gets DescriptorSetLayout from RenderPipeline. basic.vert.hlsl and basic.frag.hlsl are both PBR. GEAR_TEST updated. class Light reduced and re-designed. class Material re-designed. class Mesh loads for mesh(es) from ModelLoader and store Materials. class Model simplified.

+++ GEAR_CORE/src/Graphics/Indexbuffer.h
@@ -13,7 +13,7 @@ public:
 		void*		device;
 		void*		data;
 		size_t		size;
-		uint32_t	stride;
+		size_t		stride;
 	};
 private:
 	miru::Ref<miru::crossplatform::Buffer> m_IndexBuffer, m_IndexBufferUpload;
